Update Geometry Calculator
2.	Write a program with a switch statement that displays the following menu:
Geometry Calculator
1.	Calculate the Area of a Circle
2.	Calculate the Area of a Rectangle
3.	Calculate the Area of a Triangle
4.	Quit
If the user enters ‘1’, the program should ask for the radius of the circle and then display its area.  Use 3.14159 for PI.  
If the user enters ‘2’, the program should ask for the length and the width of the rectangle and then display its area. 
 If the user enters ‘3’, the program should ask for the length of the triangle’s base and its height and display the area.  
Entering ‘4’, should quit the application.  
Use input validation to make sure the user selects a valid menu option and does not input negative numbers for the measurements.