Create Cookie Recipe Calculator
The recipe makes 48 cookies with this amount of the ingredients. This program asks the user how many cookies he or she wants to make, and then displays the number of cups of each ingredient for the specified number of cookies.  The ingredients and amounts are in table form with 2 decimal points of precision and the decimal point always appearing.