Update Budget.cpp
This program writes to a file named “Budget.txt” that asks the user to enter the amount that he or she has budgeted for a month.  
A loop then prompts the user to enter each of his or her expenses for the month and keep a running total.  
The program writes both the expense name and corresponding expense, along with how much he or she is over-budget or under-budget to the file.
EXAMPLE:
Enter the amount for budgeting: 1200
Please enter the name of your expense followed by the amount.
Enter ‘#’ to quit.
Enter an expense: Gas 15.50
Enter an expense: Food 250.00
Enter an expense: #
Data has been written to the file.